Fix rebase and fix test
In CI, the cross-lang-lto-clang passed on x86_64 but not on aarch64.
I don’t have a good way to debug this, but with some local changes
(my clang to cross-compiling to aarch64 uses different target features,
so nothing gets inlined by default; and I wasn’t able to link a rust std
for aarch64), I think I found the culprit.

With fat lto on aarch64, the inline(never) rust function gets inlined.
This is the same behavior as with linking a C library into a Rust
binary, just that it’s different between aarch64 and x86 (fat lto on x86
inlines the noinline C function into the Rust binary but it does
not inline the inline(never) Rust function into the C binary).